Add files via upload
Three new functions for better housekeeping. 
clear_local_variables will clear all non-global variables in the workspace so that we can systematically clear stale variables.
save_all(path) will save a .mat file which contains a struct of all variables in the .m file that calls it. As well as a copy of all the .m files in that same directory so that we can know the exact state of the control code when a dataset was taken.
TestSave is a helper function used by save_all